modifies basic components structure
The app is now separated in two main components: the form and the curriculum. The form now has the role the curriculum component used to have, and in the curriculum component is where the changes made to the form will be displayed.
the 'Practical experience' section was renamed to 'Work experience'