Implements custom date input
Removed the html date input since i don't want the user to select a day, only month and year. And since the month input is not yet fully supported on all browsers, decided to make a DateInput component, wich constist in two select elements, one for years and the other for months.
The way the date is stored had to be modified for this to work.